+++ pgadmin/ctl/ctlSQLResult.cpp
@@ -354,7 +354,7 @@ wxString sqlResultTable::GetValue(int row, int col)
 				{
 					wxString data = thread->DataSet()->GetVal(col);
 
-					if (data.Length() > settings->GetMaxColSize())
+					if (data.Length() > (size_t)settings->GetMaxColSize())
 						return thread->DataSet()->GetVal(col).Left(settings->GetMaxColSize()) + wxT(" (...)");
 					else
 						return thread->DataSet()->GetVal(col);
